Moved some widget container implementations to the right package
