+++ smylib/core/src/main/java/net/smyler/smylib/gui/containers/ScrollableWidgetContainer.java
@@ -1,14 +1,12 @@
-package net.smyler.smylib.gui.screen;
+package net.smyler.smylib.gui.containers;
 
 import net.smyler.smylib.Color;
-import net.smyler.smylib.gui.containers.FlexibleWidgetContainer;
 import net.smyler.smylib.gui.widgets.Widget;
 import net.smyler.smylib.gui.widgets.ScrollbarWidget;
 import net.smyler.smylib.gui.widgets.ScrollbarWidget.ScrollbarOrientation;
 import net.smyler.smylib.gui.widgets.buttons.SpriteButtonWidget;
 import net.smyler.smylib.gui.widgets.buttons.SpriteButtonWidget.ButtonSprites;
 import net.smyler.smylib.gui.DrawContext;
-import net.smyler.smylib.gui.containers.WidgetContainer;
 
 import static net.smyler.smylib.Preconditions.checkArgument;
